+++ src/wasmtime_nif.erl
@@ -2,13 +2,19 @@
 -moduledoc false.
 
 -export([
-    compile/2,
+    compile/3,
+    validate/1,
     module_imports/1,
     module_exports/1,
     serialize/1,
     deserialize/1,
     instantiate/4,
-    call/4,
+    call/5,
+    global_get/2,
+    global_set/3,
+    table_size/2,
+    table_grow/3,
+    fuel_remaining/1,
     host_reply/3,
     interrupt/1,
     cancel/2,
@@ -20,13 +26,19 @@
     version/0
 ]).
 -nifs([
-    compile/2,
+    compile/3,
+    validate/1,
     module_imports/1,
     module_exports/1,
     serialize/1,
     deserialize/1,
     instantiate/4,
-    call/4,
+    call/5,
+    global_get/2,
+    global_set/3,
+    table_size/2,
+    table_grow/3,
+    fuel_remaining/1,
     host_reply/3,
     interrupt/1,
     cancel/2,
@@ -49,13 +61,19 @@ init() ->
         end,
     erlang:load_nif(filename:join(Priv, "wasmtime_nif"), 0).
 
-compile(_Bin, _IsWat) -> erlang:nif_error(not_loaded).
+compile(_Bin, _IsWat, _Fuel) -> erlang:nif_error(not_loaded).
+validate(_Bin) -> erlang:nif_error(not_loaded).
 module_imports(_Mod) -> erlang:nif_error(not_loaded).
 module_exports(_Mod) -> erlang:nif_error(not_loaded).
 serialize(_Mod) -> erlang:nif_error(not_loaded).
 deserialize(_Bin) -> erlang:nif_error(not_loaded).
 instantiate(_Mod, _Opts, _Ref, _Id) -> erlang:nif_error(not_loaded).
-call(_Handle, _Name, _Args, _Id) -> erlang:nif_error(not_loaded).
+call(_Handle, _Name, _Args, _Id, _Fuel) -> erlang:nif_error(not_loaded).
+global_get(_Handle, _Name) -> erlang:nif_error(not_loaded).
+global_set(_Handle, _Name, _Value) -> erlang:nif_error(not_loaded).
+table_size(_Handle, _Name) -> erlang:nif_error(not_loaded).
+table_grow(_Handle, _Name, _Delta) -> erlang:nif_error(not_loaded).
+fuel_remaining(_Handle) -> erlang:nif_error(not_loaded).
 host_reply(_Handle, _Id, _Reply) -> erlang:nif_error(not_loaded).
 interrupt(_Handle) -> erlang:nif_error(not_loaded).
 cancel(_Handle, _Id) -> erlang:nif_error(not_loaded).